+++ src/main/kotlin/mobi/hsz/idea/gitignore/CommonRunnableListeners.kt
@@ -1,4 +1,4 @@
-// Copyright 2000-2020 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+// Copyright 2000-2021 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 package mobi.hsz.idea.gitignore
 
 import com.intellij.openapi.module.Module
@@ -6,7 +6,6 @@ import com.intellij.openapi.project.ModuleListener
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.roots.ModuleRootEvent
 import com.intellij.openapi.roots.ModuleRootListener
-import com.intellij.util.Function
 import mobi.hsz.idea.gitignore.IgnoreManager.RefreshStatusesListener
 
 /**
@@ -25,6 +24,4 @@ class CommonRunnableListeners(private val task: Runnable) : RefreshStatusesListe
     override fun beforeModuleRemoved(project: Project, module: Module) = Unit
 
     override fun moduleRemoved(project: Project, module: Module) = task.run()
-
-    override fun modulesRenamed(project: Project, modules: List<Module>, oldNameProvider: Function<Module, String>) = task.run()
 }