+++ src/main/kotlin/mobi/hsz/idea/gitignore/lang/kind/CloudFoundryLanguage.kt
@@ -1,4 +1,4 @@
-// Copyright 2000-2020 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+// Copyright 2000-2021 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 package mobi.hsz.idea.gitignore.lang.kind
 
 import mobi.hsz.idea.gitignore.file.type.kind.CloudFoundryFileType
@@ -16,7 +16,4 @@ class CloudFoundryLanguage private constructor() : IgnoreLanguage("CloudFoundry"
 
     override val fileType
         get() = CloudFoundryFileType.INSTANCE
-
-    override val isVCS
-        get() = false
 }