+++ src/main/kotlin/mobi/hsz/idea/gitignore/util/Icons.kt
@@ -1,4 +1,4 @@
-// Copyright 2000-2020 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+// Copyright 2000-2021 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 package mobi.hsz.idea.gitignore.util
 
 import com.intellij.openapi.util.IconLoader
@@ -26,8 +26,14 @@ object Icons {
     /** Darcs icon. */
     val DARCS = IconLoader.getIcon("/icons/icon_darcs.png", Icons::class.java)
 
+    /** DeployHQ icon. */
+    val DEPLOYHQ = IconLoader.getIcon("/icons/icon_deployhq.png", Icons::class.java)
+
     /** Dockerignore icon. */
     val DOCKER = IconLoader.getIcon("/icons/icon_docker.png", Icons::class.java)
+
+    /** Eleventy icon. */
+    val ELEVENTY = IconLoader.getIcon("/icons/icon_eleventy.png", Icons::class.java)
 
     /** ESLint icon. */
     val ESLINT = IconLoader.getIcon("/icons/icon_eslint.png", Icons::class.java)
@@ -89,6 +95,15 @@ object Icons {
     /** TeamFoundation icon. */
     val TF = IconLoader.getIcon("/icons/icon_tf.png", Icons::class.java)
 
+    /** Tokei icon. */
+    val TOKEI = IconLoader.getIcon("/icons/icon_tokei.png", Icons::class.java)
+
     /** Up icon. */
     val UP = IconLoader.getIcon("/icons/icon_up.png", Icons::class.java)
+
+    /** Vercel icon. */
+    val VERCEL = IconLoader.getIcon("/icons/icon_vercel.png", Icons::class.java)
+
+    /** Vercel icon. */
+    val YARN = IconLoader.getIcon("/icons/icon_yarn.png", Icons::class.java)
 }